Initial Bulk Commit
Contains the initial Bulk Commit that establishes the file structure, planned files, and a basic proof of concept. The next commit will have changes with regards to content of some files but the structure and concept should remain the same. The high level concept is to have a file that is added to and removed from in parallel to a leaflet map. On export this file is export rather than the map being interacted with. This commit also may contain temporary files that are a result of exploration of possible implementations.